requirements.txt
cython~=0.29.30
fsspec~=2022.3.0
numpy~=1.22
lightgbm~=3.2.1
pandas~=1.3.5
pyspark~=3.1.3
pyyaml~=6.0
scipy~=1.7.3
scikit-learn~=1.0.2

pytest~=7.1.2